quiet a new warning that recently appeared
[warning]NetHack\src\hack.c(3223,16): Warning C4389: '!=': signed/unsigned mismatch

+++ src/hack.c
@@ -3220,7 +3220,7 @@ switch_terrain(void)
 void
 set_uinwater(int in_out)
 {
-    if (in_out != u.uinwater) {
+    if (in_out != (int) u.uinwater) {
         u.uinwater = in_out ? 1 : 0;
         switch_terrain();
     }
